Renamed OutlineProvider to DocumentSymbolProvider
